requirements.txt
Django
django-polymorphic
djangorestframework
djangorestframework-simplejwt
pyuploadcare
Pillow
django-anymail
python-dotenv
gunicorn
httpx
whitenoise
textblob
cloudinary